move flags to idiomatic serializer files
