--- dart/packages/fory/lib/src/serializer/map_flags.dart
@@ -1,27 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-abstract final class MapFlags {
-  static const int trackingKeyRef = 0x01;
-  static const int keyHasNull = 0x02;
-  static const int keyDeclaredType = 0x04;
-  static const int trackingValueRef = 0x08;
-  static const int valueHasNull = 0x10;
-  static const int valueDeclaredType = 0x20;
-}
